Initial implementation for a terminal IO system.
It's as simple as:

```
print("<red>str</red>")
```

It accepts markdown and HTML, and renders for the terminal with, if
available.

There's a LONG list of stuff that is not yet done. It's JUST good
enough to be able to get what I want for `chalk` docs.

Major things not done:

- Lots of elements don't style yet, and as a result you can't set a universal
  background color.

- Most table options aren't yet respected.

There are going to be some render bugs for a while too!

The old janky formatting crap will stick around until I have time to
fully migrate everything currently using it.