fixing http sink by not requiring ssl context
default function signature for newHttpClient is:

```
proc newHttpClient(userAgent = defUserAgent; maxRedirects = 5;
                   sslContext = getDefaultSSL(); proxy: Proxy = nil;
                   timeout = -1; headers = newHttpHeaders()): HttpClient {
```

and so when sslContext is not provided, it would default to
`getDefaultSSL()` which would require TLS certs to be installed on the
system. By providing it as nil, it bypasses all TLS logic and allow
to send report so http sinks without TLS certs installed on the system.