Jtv/next (#29)
1. Added a generic "switchboard" construct that allows for incredibly flexible I/O multiplexing, without using threads in its implementation.
2. Added a new, more subprocess IO library on top of the switchboard (which I then used to implement session capture/replay and expect-like functionality in https://github.com/crashappsec/cap10).
3. Added a one-shot, swiss-army knife subprocess interface on top of the IO library, runCommand(), which can (all optionally and in any combination) pass strings to stdin, use a pty, pass through IO between the terminal and subprocess, capture process output into strings, and more. A few things in the underlying API do require you to use the raw interface, including handling of server socket connection acceptance, using custom callbacks, and manually calling the polling loop.
4. Re-implemented the existing subprocess helpers on top of the above. For the sake of Chalk, the patchwork of similar functions are all still there, but just with one underlying implementation. This includes: runCommandGetEverything(), runInteractiveCmd(), runPager()., and more.
5. getMyIpV4Addr() Which does what it says.
6. Some tweaks to the new rendering stuff.

Note that the current implementation uses select() under the hood, instead of epoll() / kqueue(). This was partially to KISS (keep it simple), particularly since we don't currently intend to use this for massive volumes of file descriptors. They'd require two implementations that are both more complex. But I'd like to get to it at some point.

Additionally, there are some loose ends that I don't have time to fix at the moment. A small amount of the API is not directly exposed directly to nim (mainly because I'm not using it), and there's a tiny bit of cleanup that isn't properly being handled (would be a very small leak). And organizationally, I'd like to move some things around just to satisfy my OCD.