fix: always passing ssl context to request
When providing https:// url which internally redirects to https://
such as in nginx/alb/etc, in those cases if ssl context is not provided
nim segfaults vs raising exception.
To avoid segfault we pass ssl context in all cases however still
honor disallowHttp flag when the scheme is explicitly http.