refactor: improve robustness and clarity of Nim imports (#63)
Refactor imports using the process from the corresponding chalk
commit [1], with the same rationale. See that commit message for more
details.

But in particular for this repo, there was previously increased
potential for confusion from:

    import random

or

    import net

which imported nimutils modules with the same names as Nim standard
library modules [2][3]. There was also:

    import sha

which may have been confused with the Nim stdlib modules named sha1,
sha2, and sha3.

I personally find consistent import style to be helpful, and diffs like
this:

    -import unicode, options, unicodeid, unicodedb/properties, misc, strutils,
    -       random, hexdump, tables
    +import std/[unicode, options, strutils, tables]
    +import pkg/[unicodedb/properties]
    +import "."/[unicodeid, misc, random, hexdump]

and this:

    -import streams, tables, options, os, strutils, std/[net, uri, httpclient],
    -       s3client, pubsub, misc, random, encodings, std/tempfiles,
    -       parseutils, file, std/asyncfutures, net
    +import std/[streams, tables, options, os, strutils, net, uri, httpclient,
    +            tempfiles, parseutils, asyncfutures]
    +import "."/[s3client, pubsub, misc, random, encodings, file, net]

to be a significant improvement in readability. Note also the latter
imports both std/net and "."/net.

[1] https://github.com/crashappsec/chalk/commit/7fdaf52a2766
[2] https://github.com/nim-lang/Nim/blob/v2.0.0/lib/pure/net.nim
[3] https://github.com/nim-lang/Nim/blob/v2.0.0/lib/pure/random.nim