fix: resolve compiler errors from Nim 2.0.8 (#70)
Before this commit, running `nimble build` in nimutils with Nim 2.0.8
would produce compiler errors of the form:

    /foo/nimutils/nimutils/sha.nim(29, 1) Error: signature for '=destroy' must be proc[T: object](x: var T)

and one error for:

    /foo/nimutils/nimutils/advisory_lock.nim(38, 23) Error: cannot convert -1 to uint64

Resolve those errors. The former errors occur only with the `refc`
memory management strategy, which is opted into by this repo's
`nimscript.nim`:

    template applyCommonLinkOptions*(staticLink = true, quiet = true) =
      [...]
      switch("gc", "refc")
      [...]

Note however that with the `orc` memory management strategy (which is
the default from Nim 2.0.0 onwards), this commit introduces deprecation
warnings of the form:

    /foo/nimutils/nimutils/sha.nim(29, 1) Warning: A custom '=destroy' hook which takes a 'var T' parameter is deprecated; it should take a 'T' parameter [Deprecated]

It'd be possible to use conditional compilation such that the `var` is
only present for refc. But let's not consider `orc` support for now, and
remove `var` again if/when we move to `orc`.

This commit is necessary due to the below diff to compiler/semstmts.nim
between Nim 2.0.0 and 2.0.8 [1][2]:

         of attachedDestructor:
    -      localError(c.config, n.info, errGenerated,
    -        "signature for '=destroy' must be proc[T: object](x: var T) or proc[T: object](x: T)")
    +      if c.config.selectedGC in {gcArc, gcAtomicArc, gcOrc}:
    +        localError(c.config, n.info, errGenerated,
    +          "signature for '=destroy' must be proc[T: object](x: var T) or proc[T: object](x: T)")
    +      else:
    +        localError(c.config, n.info, errGenerated,
    +          "signature for '=destroy' must be proc[T: object](x: var T)")

and a change to make hex literals consistent with other literals and
the spec. From the Nim 2.0.0 Manual section on numeric literals [3]:

    The suffix starting with an apostrophe (`'`) is called a type suffix.
    Literals without a type suffix are of an integer type unless the
    literal contains a dot or `E|e` in which case it is of type float.
    This integer type is `int` if the literal is in the range
    `low(int32)..high(int32)`, otherwise it is `int64`.

In Nim 2.0.0, this worked:

    var x: uint64 = 0xffffffffffffffff
    doAssert x == 18446744073709551615'u64
    doAssert x == high(uint64)

but in Nim 2.0.8 (and Nim devel) the first line requires the `'u64` like
the second line already did previously.

[1] https://github.com/nim-lang/Nim/blob/v2.0.0/compiler/semstmts.nim#L1917-L1919
[2] https://github.com/nim-lang/Nim/blob/v2.0.8/compiler/semstmts.nim#L1968-L1974
[3] https://nim-lang.org/2.0.0/manual.html#lexical-analysis-numeric-literals